Move mock items to a separate file
