+++ frontend/constants.js
@@ -1,0 +1,70 @@
+const items = [
+  {
+    id: 0,
+    type: 'tipo',
+    name: 'nome01',
+    description: 'descrição',
+  },
+  {
+    id: 1,
+    type: 'tipo',
+    name: 'nome02',
+    description: 'descrição',
+  },
+  {
+    id: 2,
+    type: 'tipo',
+    name: 'nome03',
+    description: 'descrição',
+  },
+  {
+    id: 3,
+    type: 'tipo',
+    name: 'nome04',
+    description: 'descrição',
+  },
+  {
+    id: 4,
+    type: 'tipo',
+    name: 'nome05',
+    description: 'descrição',
+  },
+  {
+    id: 5,
+    type: 'tipo',
+    name: 'nome06',
+    description: 'descrição',
+  },
+  {
+    id: 6,
+    type: 'tipo',
+    name: 'nome07',
+    description: 'descrição',
+  },
+  {
+    id: 7,
+    type: 'tipo',
+    name: 'nome08',
+    description: 'descrição',
+  },
+  {
+    id: 8,
+    type: 'tipo',
+    name: 'nome09',
+    description: 'descrição',
+  },
+  {
+    id: 9,
+    type: 'tipo',
+    name: 'nome10',
+    description: 'descrição',
+  },
+  {
+    id: 10,
+    type: 'tipo',
+    name: 'nome11',
+    description: 'descrição',
+  },
+];
+
+export default items;
