+++ frontend/pages/index.js
@@ -2,44 +2,7 @@ import Head from 'next/head';
 import ItemList from '../components/ItemList';
 import styles from '../styles/Home.module.css';
 
-const items = [
-  {
-    id: 0,
-    type: 'tipo',
-    name: 'nome01',
-    description: 'descrição',
-  },
-  {
-    id: 1,
-    type: 'tipo',
-    name: 'nome02',
-    description: 'descrição',
-  },
-  {
-    id: 2,
-    type: 'tipo',
-    name: 'nome03',
-    description: 'descrição',
-  },
-  {
-    id: 3,
-    type: 'tipo',
-    name: 'nome04',
-    description: 'descrição',
-  },
-  {
-    id: 4,
-    type: 'tipo',
-    name: 'nome05',
-    description: 'descrição',
-  },
-  {
-    id: 5,
-    type: 'tipo',
-    name: 'nome06',
-    description: 'descrição',
-  },
-];
+import items from '../constants';
 
 export default function Home() {
   return (
